Issue #31: Removed flask-jwt dependency
Instead of converting the extension to use flask-jwt-extended instead of
flask-jwt, I just moved to dependence on pyjwt directly. The
flask-jwt-exteneded extensions is very opinionated and seemed to require
some support that I didn't feel was necessary in this extension. Moving
to using pyjwt directly made the package a lot slimmer in parts, and
gave a lot more control of the package.

The package now supports token refreshing, but implementing the
blacklist and checking is left to the user. Now, a token has both a
lifespan for being accepted. After that is expired, it may be refreshed
within a refresh-lifespan to get a new token issued without
re-authenticating with username and password.

As before, the implementation of the user class is left to the client
code, but there are a few very minimal requirements.