Issue 115: Added snippets to tutorial
Used literal includes for this with specific line numbers, so it will be
fun keeping these aligned with the source code. Still, it's very nice

Also changed the structure a bit. Got rid of the curl tuorial since it
seemed redundant and kinda confusing. I figure if someon wants to use a
tool like curl or postman, they will be savvy enough to follow along and
structure their requests as needed