Issue #135: Made auth_required optional with other decorators
It used to be that you had to explicitly declare 'auth_required' when
using the other decorators. The order was also important. This was
unnecessarily cumbersome, and it was confusing when declaring dynamic
routes with flask_restplus as you need to add decorators in a list. The
order of said list had to be reverse of how they are specified as
decorators in statically declared routes.

The easiest fix for this was to have the decorators implicitly check for
athorization themselves if the app context didn't already have jwt data.

The result of this is that you may or may not include an explicit
@auth_required when you are using one of the other decorators.