Issue #315: Added support for UUID primary keys
Updated the `encode_jwt_token()` method so that it can use non-JSON
serializable `id` fields from the user class. This includes UUID primary
keys. If the `id` field is JSON serializable, it will be serialized
as-is. If it is not JSON serializable, it will be converted to a string
first.